+++ runtime/src/lib.rs
@@ -78,56 +78,6 @@ stability_scope!(BETA {
     /// Default [`Blob`] version used when no version is specified via [`Storage::open`].
     pub const DEFAULT_BLOB_VERSION: u16 = 0;
 
-    /// Version of a [`Blob`]'s on-disk header layout.
-    ///
-    /// This versions the runtime's on-disk container (where data begins), not the blob's
-    /// contents: the application-owned blob version passed to [`Storage::open_versioned`] is a
-    /// separate field and is unaffected by the layout.
-    ///
-    /// New blobs are always created with the latest layout. Reopening an existing blob honors
-    /// the layout recorded in its header.
-    #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-    pub(crate) enum BlobHeaderLayout {
-        /// An 8-byte header, with data beginning immediately after it.
-        V0,
-        /// A header padded to one 4096-byte page, so data begins on an aligned boundary.
-        V1,
-    }
-
-    impl BlobHeaderLayout {
-        /// The runtime version recorded in a header of this layout.
-        pub(crate) const fn runtime_version(self) -> u16 {
-            match self {
-                Self::V0 => 0,
-                Self::V1 => 1,
-            }
-        }
-
-        /// The magic bytes recorded in a header of this layout: a fixed 3-byte brand (`CWI`,
-        /// "is this file ours?") followed by a 1-byte layout tag ("which container layout?").
-        ///
-        /// The layout tag lives in the magic rather than the runtime version field because V0
-        /// stamped that field as zero, and zeros are exactly what a torn header write leaves
-        /// behind. Tags are nonzero and distinct, so no layout's magic can be turned into
-        /// another's by zeroing bytes, and a torn write can never be misread as a complete
-        /// header of a different layout.
-        pub(crate) const fn magic(self) -> [u8; 4] {
-            match self {
-                Self::V0 => *b"CWIC", // Commonware Is CWIC
-                Self::V1 => *b"CWI1",
-            }
-        }
-
-        /// The layout recorded by a header with the given magic bytes, if supported.
-        pub(crate) const fn from_magic(magic: &[u8; 4]) -> Option<Self> {
-            match magic {
-                b"CWIC" => Some(Self::V0),
-                b"CWI1" => Some(Self::V1),
-                _ => None,
-            }
-        }
-    }
-
     /// Errors that can occur when interacting with the runtime.
     #[derive(Error, Debug, Clone)]
     pub enum Error {
