+++ runtime/src/storage/memory.rs
@@ -269,7 +269,8 @@ impl crate::Blob for Blob {
 mod tests {
     use super::{Header, *};
     use crate::{
-        Blob, BlobHeaderLayout, BufferPoolConfig, Storage as _, storage::tests::run_storage_tests,
+        Blob, BufferPoolConfig, Storage as _,
+        storage::{BlobHeaderLayout, tests::run_storage_tests},
         telemetry::metrics::Registry,
     };
